requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.10.5
colorama==0.4.6
distro==1.9.0
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
openai==2.6.1
pillow==11.3.0
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
wheel==0.45.1
